Initial JWT implementation

+++ backend/src/controllers/consumerController.js
@@ -12,6 +12,7 @@
  *    George Benos (Telesto Technologies)
  */
 import consumerService from "../services/consumerService"
+import { id_token } from ".."
 
 export class consumerController{
 
@@ -31,6 +32,10 @@ export class consumerController{
 
     createConsumer = async function createConsumer(req, res, next){
         const consumerObj = req.body
+        //Auth stuff: No one has access to this function (use debugging flag to access)
+        if (!id_token.debug){
+            return (res.status(403).json({error: "You are not authorized to create new users"}))
+        }
         try{  
             const consumer = await consumerService.createConsumer(consumerObj)
             return res.status(201).json({
@@ -45,6 +50,10 @@ export class consumerController{
     editConsumer = async function editConsumer(req, res, next){
         const consumerObj = req.body
         const did = req.params.did
+        //Auth stuff: A user can only edit themselves
+        if (id_token.did !== did){
+            return (res.status(403).json({error: "You are not authorized to edit this user"}))
+        }
         try{
             const consumer = await consumerService.editConsumer(did, consumerObj.email)
             if (!consumer) return res.status(404).json({error: `Consumer with did ${did} does not exist`})
@@ -59,6 +68,10 @@ export class consumerController{
 
     deleteConsumer = async function deleteConsumer(req, res, next){
         const did = req.params.did
+        //Auth stuff: No one has access to this function (use debugging flag to access)
+        if (!id_token.debug){
+            return (res.status(403).json({error: "You are not authorized to delete users"}))
+        }
         try{
             const consumer = await consumerService.deleteConsumer(did)
             if (!consumer) return res.status(404).json({error: `Consumer with did ${did} does not exist`})
