+++ backend/src/controllers/ratingController.js
@@ -13,6 +13,7 @@
  */
 
 import ratingService from "../services/ratingService"
+import { id_token } from ".."
 
 export class RatingController{
 
@@ -44,6 +45,10 @@ export class RatingController{
 
     createRating = async function createRating(req, res, next){
         const ratingObj = req.body
+        //Auth stuff: A consumer can only create a rating as himself
+        if (id_token.did !== ratingObj.byConsumer){
+            return (res.status(403).json({error: "You can only create ratings where you are consumer"}))
+        }
         try{
             const rating = await ratingService.createRating(ratingObj)
             return res.status(201).json({
@@ -60,7 +65,15 @@ export class RatingController{
         const msg = req.body.msg || ""
         const id = req.params.id
         try{
-            const rating = await ratingService.editRating(id, ratings, msg)
+            //Auth stuff: A consumer can only edit a rating of their own
+            let rating = await ratingService.getRating(id)
+            if (!rating){
+                return (res.status(404).json({error: `Rating with id=${id} does not exist`}))
+            }
+            else if (rating.byConsumer !== id_token.did){
+                return (res.status(403).json({error: "You are not authorized to edit this rating"}))
+            }
+            rating = await ratingService.editRating(id, ratings, msg)
             if (!rating) return res.status(404).json({error: `Rating with id: ${id} does not exist`})
             return res.status(200).json({
                 rating: rating
@@ -73,6 +86,10 @@ export class RatingController{
 
     deleteRating = async function deleteRating(req, res, next){
         const id = req.params.id
+        //Auth stuff: No one has access to this function (use debugging flag to access)
+        if (!id_token.debug){
+            return (res.status(403).json({error: "You are not authorized to delete ratings"}))
+        }
         try{
         const rating = await ratingService.deleteRating(id)
             if (!rating) return res.status(404).json({error: `Rating with id: ${id} does not exist`})
@@ -89,7 +106,15 @@ export class RatingController{
         const id = req.params.id
         const response = req.body.response
         try{
-            const rating = await ratingService.respondToRating(id, response)
+            //Auth stuff: A provider can only respond to a rating that refers to him
+            let rating = await ratingService.getRating(id)
+            if (!rating){
+                return (res.status(404).json({error: `Rating with id=${id} does not exist`}))
+            }
+            else if (rating.forProvider !== id_token.did){
+                return (res.status(403).json({error: "You are not authorized to respond to this rating"}))
+            }
+            rating = await ratingService.respondToRating(id, response)
             if (!rating) return res.status(422).json({error: `Rating with id: ${id} does not exist`})
             return res.status(200).json({
                 rating: rating
